+++ packages/react/src/components/crypto/AaveProtocol/index.tsx
@@ -1,0 +1,83 @@
+interface AaveProtocolProps {
+  className?: string
+  size?: number | string
+  fill?: string
+  inverted?: boolean
+  weight?: 'thin' | 'light' | 'regular' | 'bold' | 'outline' | 'fill'
+}
+
+export function AaveProtocol({
+  size = '1em',
+  className,
+  fill = 'none',
+  inverted = false,
+  weight = 'light',
+  ...rest
+}: AaveProtocolProps) {
+  let selectedWeight = ''
+  let strokeOut = 'currentColor'
+  let strokeIn = 'currentColor'
+
+  switch (weight) {
+    case 'thin':
+      selectedWeight = '1'
+      break
+    case 'light':
+      selectedWeight = '1.5'
+      break
+    case 'regular':
+      selectedWeight = '2'
+      break
+    case 'bold':
+      selectedWeight = '2.5'
+      break
+    case 'outline':
+      strokeIn = inverted ? '#1a1a1a' : '#fff'
+      selectedWeight = '1.5'
+      break
+    case 'fill':
+      strokeIn = '#fff'
+      fill = 'currentColor'
+      selectedWeight = '1.5'
+      break
+    default:
+      break
+  }
+
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size}
+      height={size}
+      fill={fill}
+      viewBox="0 0 24 24"
+      {...rest}
+    >
+      <path
+        className={className}
+        stroke={strokeOut}
+        strokeWidth={selectedWeight}
+        strokeMiterlimit="10"
+        d="M12 22C17.5228 22 22 17.5228 22 12C22 6.47715 17.5228 2 12 2C6.47715 2 2 6.47715 2 12C2 17.5228 6.47715 22 12 22Z"
+      />
+      <path
+        className={className}
+        stroke={strokeIn}
+        strokeWidth={selectedWeight}
+        strokeMiterlimit="10"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M8 16L12 7L16 16"
+      />
+      <path
+        className={className}
+        stroke={strokeIn}
+        strokeWidth={selectedWeight}
+        strokeMiterlimit="10"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M8 12H12"
+      />
+    </svg>
+  )
+}