+++ packages/react/src/components/grid-alignment/SquarePile/index.tsx
@@ -2,14 +2,45 @@ interface SquarePileProps {
   className?: string
   size?: number | string
   fill?: string
+  inverted?: boolean
+  weight?: 'thin' | 'light' | 'regular' | 'bold' | 'outline' | 'fill'
 }
 
 export function SquarePile({
   size = '1em',
   className,
   fill = 'none',
+  inverted = false,
+  weight = 'light',
   ...rest
 }: SquarePileProps) {
+  let selectedWeight = ''
+  let stroke = 'currentColor'
+
+  switch (weight) {
+    case 'thin':
+      selectedWeight = '1'
+      break
+    case 'light':
+      selectedWeight = '1.5'
+      break
+    case 'regular':
+      selectedWeight = '2'
+      break
+    case 'bold':
+      selectedWeight = '3'
+      break
+    case 'outline':
+      stroke = inverted ? '#1a1a1a' : 'currentColor'
+      selectedWeight = '1.5'
+      break
+    case 'fill':
+      fill = 'currentColor'
+      break
+    default:
+      break
+  }
+
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -21,26 +52,26 @@ export function SquarePile({
     >
       <path
         className={className}
-        stroke="currentColor"
+        stroke={stroke}
         strokeLinecap="round"
         strokeLinejoin="round"
-        strokeWidth="1.5"
+        strokeWidth={selectedWeight}
         d="M12 17v2.38c0 1.87-.75 2.62-2.63 2.62H4.62C2.75 22 2 21.25 2 19.38v-4.75C2 12.75 2.75 12 4.62 12H7v2.37C7 16.25 7.75 17 9.62 17H12z"
       ></path>
       <path
         className={className}
-        stroke="currentColor"
+        stroke={stroke}
         strokeLinecap="round"
         strokeLinejoin="round"
-        strokeWidth="1.5"
+        strokeWidth={selectedWeight}
         d="M17 12v2.37c0 1.88-.75 2.63-2.63 2.63H9.62C7.75 17 7 16.25 7 14.37V9.62C7 7.75 7.75 7 9.62 7H12v2.37c0 1.88.75 2.63 2.62 2.63H17z"
       ></path>
       <path
         className={className}
-        stroke="currentColor"
+        stroke={stroke}
         strokeLinecap="round"
         strokeLinejoin="round"
-        strokeWidth="1.5"
+        strokeWidth={selectedWeight}
         d="M22 4.62v4.75c0 1.88-.75 2.63-2.63 2.63h-4.75C12.75 12 12 11.25 12 9.37V4.62C12 2.75 12.75 2 14.62 2h4.75C21.25 2 22 2.75 22 4.62z"
       ></path>
     </svg>