+++ packages/react/src/components/social/Support24/index.tsx
@@ -1,0 +1,83 @@
+interface Support24Props {
+  className?: string
+  size?: number | string
+  fill?: string
+  inverted?: boolean
+  weight?: 'thin' | 'light' | 'regular' | 'bold' | 'outline' | 'fill'
+}
+
+export function Support24({
+  size = '1em',
+  className,
+  fill = 'none',
+  inverted = false,
+  weight = 'light',
+  ...rest
+}: Support24Props) {
+  let selectedWeight = ''
+  let strokeOut = 'currentColor'
+  let strokeIn = 'currentColor'
+
+  switch (weight) {
+    case 'thin':
+      selectedWeight = '1'
+      break
+    case 'light':
+      selectedWeight = '1.5'
+      break
+    case 'regular':
+      selectedWeight = '2'
+      break
+    case 'bold':
+      selectedWeight = '2.5'
+      break
+    case 'outline':
+      strokeIn = inverted ? '#1a1a1a' : '#fff'
+      selectedWeight = '1.5'
+      break
+    case 'fill':
+      strokeIn = '#fff'
+      fill = 'currentColor'
+      selectedWeight = '1.5'
+      break
+    default:
+      break
+  }
+
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size}
+      height={size}
+      fill={fill}
+      viewBox="0 0 24 24"
+      {...rest}
+    >
+      <path
+        className={className}
+        stroke={strokeOut}
+        strokeWidth={selectedWeight}
+        strokeMiterlimit="10"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M18 18.8597H17.24C16.44 18.8597 15.68 19.1697 15.12 19.7297L13.41 21.4198C12.63 22.1898 11.36 22.1898 10.58 21.4198L8.87 19.7297C8.31 19.1697 7.54 18.8597 6.75 18.8597H6C4.34 18.8597 3 17.5298 3 15.8898V4.97974C3 3.33974 4.34 2.00977 6 2.00977H18C19.66 2.00977 21 3.33974 21 4.97974V15.8898C21 17.5198 19.66 18.8597 18 18.8597Z"
+      />
+      <path
+        className={className}
+        stroke={strokeIn}
+        strokeWidth={selectedWeight}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M7 9.16003C7 8.23003 7.76 7.46997 8.69 7.46997C9.62 7.46997 10.38 8.23003 10.38 9.16003C10.38 11.04 7.71 11.24 7.12 13.03C7 13.4 7.31 13.77 7.7 13.77H10.38"
+      />
+      <path
+        className={className}
+        stroke={strokeIn}
+        strokeWidth={selectedWeight}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M16.04 13.76V8.05003C16.04 7.79003 15.87 7.55997 15.62 7.48997C15.37 7.41997 15.1 7.51997 14.96 7.73997C14.24 8.89997 13.46 10.22 12.78 11.38C12.67 11.57 12.67 11.82 12.78 12.01C12.89 12.2 13.1 12.3199 13.33 12.3199H17"
+      />
+    </svg>
+  )
+}