+++ packages/react/src/components/warning-check/AddSquare/index.tsx
@@ -1,0 +1,115 @@
+interface AddSquareProps {
+  className?: string
+  size?: number | string
+  fill?: string
+  inverted?: boolean
+  weight?: 'thin' | 'light' | 'regular' | 'bold' | 'outline' | 'fill'
+}
+
+export function AddSquare({
+  size = '1em',
+  className,
+  fill = 'none',
+  inverted = false,
+  weight = 'light',
+  ...rest
+}: AddSquareProps) {
+  let selectedWeight = ''
+  let strokeOut = 'currentColor'
+  let strokeIn = 'currentColor'
+
+  switch (weight) {
+    case 'thin':
+      selectedWeight = '1'
+      break
+    case 'light':
+      selectedWeight = '1.5'
+      break
+    case 'regular':
+      selectedWeight = '2'
+      break
+    case 'bold':
+      selectedWeight = '3'
+      break
+    case 'outline':
+      strokeIn = inverted ? '#1a1a1a' : '#fff'
+      selectedWeight = '1.5'
+      break
+    case 'fill':
+      strokeIn = '#fff'
+      fill = 'currentColor'
+      selectedWeight = '2'
+      break
+    default:
+      break
+  }
+
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size}
+      height={size}
+      fill={fill}
+      viewBox="0 0 24 24"
+      {...rest}
+    >
+      <path
+        className={className}
+        stroke={strokeOut}
+        strokeWidth={selectedWeight}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M9 22H15C20 22 22 20 22 15V9C22 4 20 2 15 2H9C4 2 2 4 2 9V15C2 20 4 22 9 22Z"
+      />
+      <path
+        className={className}
+        stroke={strokeIn}
+        strokeWidth={selectedWeight}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M8 12H16"
+      />
+      <path
+        className={className}
+        stroke={strokeIn}
+        strokeWidth={selectedWeight}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M12 16V8"
+      />
+    </svg>
+    // <svg
+    //   xmlns="http://www.w3.org/2000/svg"
+    //   width={size}
+    //   height={size}
+    //   fill={fill}
+    //   viewBox="0 0 24 24"
+    //   {...rest}
+    // >
+    //   <path
+    //     className={className}
+    //     stroke={strokeOut}
+    //     strokeWidth={selectedWeight}
+    //     strokeLinecap="round"
+    //     strokeLinejoin="round"
+    //     d="M12 22C17.5 22 22 17.5 22 12C22 6.5 17.5 2 12 2C6.5 2 2 6.5 2 12C2 17.5 6.5 22 12 22Z"
+    //   />
+    //   <path
+    //     className={className}
+    //     stroke={strokeIn}
+    //     strokeWidth={selectedWeight}
+    //     strokeLinecap="round"
+    //     strokeLinejoin="round"
+    //     d="M8 12H16"
+    //   />
+    //   <path
+    //     className={className}
+    //     stroke={strokeIn}
+    //     strokeWidth={selectedWeight}
+    //     strokeLinecap="round"
+    //     strokeLinejoin="round"
+    //     d="M12 16V8"
+    //   />
+    // </svg>
+  )
+}