+++ packages/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @uicons/react
 
+## 1.1.0
+
+### Minor Changes
+
+- Add weight and inverted properties on component
+
+## 1.0.5
+
+### Patch Changes
+
+- Separating the components into category
+
+## 1.0.4
+
+### Patch Changes
+
+- Add 2 new icon components
+
 ## 1.0.3
 
 ### Patch Changes